Fetch data dynamically from API
Added an API client to use OMDB for fetching movie details when adding
a new movie rating to provide a better user experience and make the
process of adding new movies more convenient and failsafe regarding
different spellings and variations in punctuation within a movie title.

Finally, the changes also allow a movie to have multiple countries of
origin, implementing the recent upgrade of the database schema.

The new functionality has been completely implemented in the
'Add movie rating' process which can be considered as a blueprint for
the remaining CLI commands / processes which still need to be tied
to API and database calls.