Tie remaining CLI commands to API and DB calls
Updated all CLI menu commands to retrieve data from the database
or in the case of not already stored movies, fetch the data
via the OMDB API (adding new movies).

Removed obsolete 'movies.json'.

Additionally, for all commands relying on (the exact)
movie title in the database, fuzzy search is used, providing
the option to let the user choose the movie title
an action should be performed on.

- NB -
To address the following issue, an urgent update
needs to be implemented ASAP:

As there are cases of multiple movies with the same title,
further adjustments to the database schema, queries and code
are needed to allow for the identification of a movie,
by its imdbID instead of its title. Additionally country flag
emojis will be used to let the user visually discriminate between
same title, same year movies.